+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ## 0.0.1
 ### 2018-10-01
 - Initial version
+- Configuration to be able to send email using a non-embedded SMTP server